+++ examples/topology/3dBorderExtractionImg.cpp
@@ -49,7 +49,7 @@
 #include "DGtal/io/readers/VolReader.h"
 #include "DGtal/images/ImageSelector.h"
 #include "DGtal/images/imagesSetsUtils/SetFromImage.h"
-#include "DGtal/io/viewers/Viewer3D.h"
+#include "DGtal/io/viewers/PolyscopeViewer.h"
 
 #include "DGtal/helpers/StdDefs.h"
 #include "ConfigExamples.h"
@@ -66,9 +66,7 @@ int main( int argc, char** argv )
 {
 
   std::string inputFilename = examplesPath + "samples/Al.100.vol";
-  QApplication application(argc,argv);
-  Viewer3D<> viewer;
-  viewer.show();
+  PolyscopeViewer<> viewer;
 
   typedef ImageSelector < Z3i::Domain, int>::Type Image;
   Image image = VolReader<Image>::importVol(inputFilename);
@@ -77,9 +75,10 @@ int main( int argc, char** argv )
   Z3i::Object18_6 obj3d (Z3i::dt18_6, set3d);
   Z3i::Object18_6 border = obj3d.border();
   viewer << border;
-  viewer << ClippingPlane(0,1,0, -40) << Viewer3D<>::updateDisplay;
+  viewer << ClippingPlane(0,1,0, -40);
 
-  return application.exec();
+  viewer.show();
+  return 0;
 }
 //                                                                           //
 ///////////////////////////////////////////////////////////////////////////////