+++ src/DGtal/helpers/ShortcutsGeometry.h
@@ -48,6 +48,7 @@
 #include "DGtal/geometry/surfaces/estimation/IIGeometricFunctors.h"
 #include "DGtal/geometry/surfaces/estimation/IntegralInvariantVolumeEstimator.h"
 #include "DGtal/geometry/surfaces/estimation/IntegralInvariantCovarianceEstimator.h"
+#include "DGtal/geometry/meshes/CorrectedNormalCurrentComputer.h"
 
 #include "DGtal/dec/DiscreteExteriorCalculusFactory.h"
 #include "DGtal/dec/ATSolver2D.h"
@@ -169,6 +170,7 @@ namespace DGtal
       typedef typename functors::IIPrincipalCurvaturesAndDirectionsFunctor<Space>::Quantity   CurvatureTensorQuantity;
       typedef std::vector< CurvatureTensorQuantity >              CurvatureTensorQuantities;
 
+      typedef CorrectedNormalCurrentComputer<RealPoint, RealVector> CNCComputer;
 
       typedef TrueDigitalSurfaceLocalEstimator
         < KSpace, ImplicitShape3D, PositionFunctor >                TruePositionEstimator;
@@ -220,7 +222,7 @@ namespace DGtal
       static Parameters parametersShapeGeometry()
       {
         return Parameters
-          ( "projectionMaxIter", 20 )
+          ( "projectionMaxIter",  20 )
           ( "projectionAccuracy", 0.0001 )
           ( "projectionGamma",    0.5 )
           ( "gridstep",           1.0 );
@@ -372,6 +374,106 @@ namespace DGtal
         return n_true_estimations;
       }
 
+    /// Given a SurfaceMesh, compute mean curvature at each face using 
+    /// CorrectedNormalCurrent method.
+    ///
+    /// @warning In this code, only triangles with barycenters strictly inside the sphere are 
+    /// considered.
+    ///
+    /// @param mesh The surface mesh
+    /// @param faces The faces to compute curvature at
+    /// @param params
+    ///   - unit_u: Whether the computed normals should be normalized or not
+    ///   - r-radius        [   3.0]: the constant for kernel radius parameter r in r(h)=r h^alpha (VCM,II,Trivial).
+    ///   - alpha           [  0.33]: the parameter alpha in r(h)=r h^alpha (VCM, II, CNC)."
+    ///   - gridstep        [   1.0]: the digitization gridstep (often denoted by h).
+    /// @return The curvatures for each face of the mesh, in the same order `faces` 
+    static Scalars
+      getCNCMeanCurvatures
+      ( CountedPtr<typename Base::SurfaceMesh>  mesh, 
+        const typename Base::SurfaceMesh::Faces faces, 
+        const Parameters&                       params = parametersShapeGeometry() )
+      {
+        using Face = typename Base::SurfaceMesh::Face;
+
+        bool unit_u = params["unit_u"].as<int>();
+        double radius = params["r-radius"].as<double>();
+        double alpha  = params["alpha"].as<double>();
+        double h      = params["gridstep"].as<double>();
+        if ( alpha != 1.0 ) radius *= pow( h, alpha-1.0 );
+
+        CNCComputer computer(*mesh, unit_u);
+        
+        const auto& mu0 = computer.computeMu0();
+        const auto& mu1 = computer.computeMu1();
+        
+        Scalars curvatures(faces.size());
+        for (size_t i = 0; i < faces.size(); ++i)
+        {
+          const auto center = mesh->faceCentroid(faces[i]);
+          const auto area = mu0.measure(center, radius, faces[i]);
+          const auto lmu1 = mu1.measure(center, radius, faces[i]);
+          curvatures[i] = CNCComputer::meanCurvature(area, lmu1);
+        }
+
+        return curvatures;
+      }
+
+      /// Given a SurfaceMesh, compute mean curvature at each face using
+      /// CorrectedNormalCurrent method.
+      ///
+      /// This overloads compute curvature for each face of the mesh.
+      ///
+      /// @warning In this code, only triangles with barycenters
+      /// strictly inside the sphere are considered.
+      ///
+      /// @param mesh The surface mesh
+      /// @param params
+      ///   - unit_u: Whether the computed normals should be normalized or not
+      ///   - r-radius        [   3.0]: the constant for kernel radius parameter
+      ///   r in r(h)=r h^alpha (VCM,II,Trivial).
+      ///   - alpha           [  0.33]: the parameter alpha in r(h)=r h^alpha
+      ///   (VCM, II, CNC)."
+      ///   - gridstep        [   1.0]: the digitization gridstep (often denoted
+      ///   by h).
+      /// @return The curvatures for each face of mesh, in the the order given by
+      /// the mesh
+      static Scalars getCNCMeanCurvatures(
+      CountedPtr<typename Base::SurfaceMesh> mesh,
+      const Parameters & params = parametersShapeGeometry() )
+      {
+        std::vector<typename Base::SurfaceMesh::Face> allFaces(mesh->nbFaces());
+        std::iota(allFaces.begin(), allFaces.end(), 0);
+
+        return getCNCMeanCurvatures(mesh, allFaces, params);
+      }
+
+      /// Given a SurfaceMesh, compute mean curvature at each face using
+      /// CorrectedNormalCurrent method.
+      ///
+      /// @warning In this code, only triangles with barycenters strictly inside
+      /// the sphere are considered.
+      ///
+      /// @tparam Any digital object convertible to surface mesh via
+      /// Shortcuts::makePrimalSurfaceMesh
+      /// @param digitalObject A digital object
+      /// @param params
+      ///   - unit_u: Whether the computed normals should be normalized or not
+      ///   - r-radius        [   3.0]: the constant for kernel radius parameter
+      ///   r in r(h)=r h^alpha (VCM,II,Trivial).
+      ///   - alpha           [  0.33]: the parameter alpha in r(h)=r h^alpha
+      ///   (VCM, II, CNC)."
+      ///   - gridstep        [   1.0]: the digitization gridstep (often denoted
+      ///   by h).
+      /// @return The curvatures for each face of the triangulated surface object
+      template <typename T>
+      static Scalars getCNCMeanCurvatures(
+      T & digitalObject, const Parameters & params = parametersShapeGeometry() )
+      {
+        CountedPtr<typename Base::SurfaceMesh> mesh = Base::makePrimalSurfaceMesh(digitalObject);
+        return getCNCMeanCurvatures(mesh, params);
+      }
+
       /// Given a space \a K, an implicit \a shape, a sequence of \a
       /// surfels, and a gridstep \a h, returns the gaussian curvatures at the
       /// specified surfels, in the same order.
@@ -412,10 +514,111 @@ namespace DGtal
         return n_true_estimations;
       }
 
-      /// Given a space \a K, an implicit \a shape, a sequence of \a
-      /// surfels, and a gridstep \a h, returns the first (smallest)
-      /// principal curvatures at the specified surfels, in the same
-      /// order.
+      /// Given a SurfaceMesh, compute gaussian curvature at each face using
+      /// CorrectedNormalCurrent method.
+      ///
+      /// @warning  In this code, only triangles with barycenters strictly
+      /// inside the sphere are considered.
+      ///
+      /// @param mesh The surface mesh
+      /// @param faces The faces to compute curvature at
+      /// @param params
+      ///   - unit_u: Whether the computed normals should be normalized or not
+      ///   - r-radius        [   3.0]: the constant for kernel radius parameter
+      ///   r in r(h)=r h^alpha (VCM,II,Trivial).
+      ///   - alpha           [  0.33]: the parameter alpha in r(h)=r h^alpha
+      ///   (VCM, II, CNC)."
+      ///   - gridstep        [   1.0]: the digitization gridstep (often denoted
+      ///   by h).
+      /// @return The curvatures for each face of the mesh, in the same order
+      /// `faces`
+      static Scalars getCNCGaussianCurvatures(
+      CountedPtr<typename Base::SurfaceMesh> mesh,
+      const typename Base::SurfaceMesh::Faces & faces,
+      const Parameters & params = parametersShapeGeometry() )
+      {
+        using Face = typename Base::SurfaceMesh::Face;
+
+        bool unit_u = params["unit_u"].as<int>();
+        double radius = params["r-radius"].as<double>();
+        double alpha  = params["alpha"].as<double>();
+        double h      = params["gridstep"].as<double>();
+        if ( alpha != 1.0 ) radius *= pow( h, alpha-1.0 );
+
+        CNCComputer computer(*mesh, unit_u);
+        
+        const auto& mu0 = computer.computeMu0();
+        const auto& mu2 = computer.computeMu2();
+
+        Scalars curvatures(faces.size());
+        for (size_t i = 0; i < faces.size(); ++i)
+        {
+          const auto center = mesh->faceCentroid(faces[i]);
+          const auto area = mu0.measure(center, radius, faces[i]);
+          const auto lmu2 = mu2.measure(center, radius, faces[i]);
+          curvatures[i] = CNCComputer::GaussianCurvature(area, lmu2);
+        }
+        
+        return curvatures;
+      }
+
+      /// Given a SurfaceMesh, compute gaussian curvature at each face using
+      /// CorrectedNormalCurrent method.
+      ///
+      /// This overloads compute curvature for each face of the mesh.
+      ///
+      /// @warning  In this code, only triangles with barycenters strictly
+      /// inside the sphere are considered.
+      ///
+      /// @param mesh The surface mesh
+      /// @param params
+      ///   - unit_u: Whether the computed normals should be normalized or not
+      ///   - r-radius        [   3.0]: the constant for kernel radius parameter
+      ///   r in r(h)=r h^alpha (VCM,II,Trivial).
+      ///   - alpha           [  0.33]: the parameter alpha in r(h)=r h^alpha
+      ///   (VCM, II, CNC)."
+      ///   - gridstep        [   1.0]: the digitization gridstep (often denoted
+      ///   by h).
+      /// @return The curvatures for each face of the mesh, in thn the order
+      /// given by the mesh
+      static Scalars getCNCGaussianCurvatures(
+      CountedPtr<typename Base::SurfaceMesh> mesh,
+      const Parameters & params = parametersShapeGeometry() )
+      {
+        std::vector<typename Base::SurfaceMesh::Face> allFaces(mesh->nbFaces());
+        std::iota(allFaces.begin(), allFaces.end(), 0);
+        
+        return getCNCGaussianCurvatures(mesh, allFaces, params);
+      }
+
+      /// Given a SurfaceMesh, compute mean curvature at each face using
+      /// CorrectedNormalCurrent method.
+      ///
+      /// @warning  In this code, only triangles with barycenters strictly
+      /// inside the sphere are considered.
+      ///
+      /// @tparam T Any digital object convertible to surface mesh via
+      /// Shortcuts::makePrimalSurfaceMesh
+      /// @param digitalObject A digital object
+      /// @param params
+      ///   - unit_u: Whether the computed normals should be normalized or not
+      ///   - r-radius        [   3.0]: the constant for kernel radius parameter
+      ///   r in r(h)=r h^alpha (VCM,II,Trivial).
+      ///   - alpha           [  0.33]: the parameter alpha in r(h)=r h^alpha
+      ///   (VCM, II, CNC)."
+      ///   - gridstep        [   1.0]: the digitization gridstep (often denoted
+      ///   by h).
+      /// @return The curvatures for each face of the triangulated surface object
+      template <typename T>
+      static Scalars getCNCGaussianCurvatures(
+      T & digitalObject, const Parameters & params = parametersShapeGeometry() )
+      {
+        CountedPtr<typename Base::SurfaceMesh> mesh = Base::makePrimalSurfaceMesh(digitalObject);
+        return getCNCGaussianCurvatures(mesh, params);
+      }
+
+      /// Given a space \a K, an implicit \a shape, a sequence of \a principal curvatures 
+      /// at the specified surfels, in the same order.
       ///
       /// @note that the first principal curvature is approximated by projecting the
       /// surfel centroid onto the implicit 3D shape.
@@ -623,7 +826,140 @@ namespace DGtal
         return n_true_estimations;
       }
 
-      /// @}
+      /// Given a SurfaceMesh, compute principal curvature at each face using
+      /// CorrectedNormalCurrent method.
+      ///
+      /// @note If no normals are provided for the faces, the normals will be
+      /// computed (and set) using vertex normals if they exist and positions
+      /// otherwise.
+      ///
+      /// @warning  In this code, only triangles with barycenters strictly
+      /// inside the sphere are considered.
+      ///
+      /// @param[in,out] mesh The surface mesh. The mesh will be modified if no
+      /// face normals are provided.
+      /// @param[in] faces The faces to compute curvature at
+      /// @param[in] params
+      ///   - unit_u: Whether the computed normals should be normalized or not
+      ///   - r-radius        [   3.0]: the constant for kernel radius parameter
+      ///   r in r(h)=r h^alpha (VCM,II,Trivial).
+      ///   - alpha           [  0.33]: the parameter alpha in r(h)=r h^alpha
+      ///   (VCM, II, CNC)."
+      ///   - gridstep        [   1.0]: the digitization gridstep (often denoted
+      ///   by h).
+      /// @return The principal curvatures for each face of the mesh, in the same
+      /// order as faces. The result is a 4-element tuples: [first curvatures,
+      /// second curvatures, first directions, second directions].
+      static std::tuple<Scalars, Scalars, RealVectors, RealVectors>
+        getCNCPrincipalCurvaturesAndDirections
+        ( CountedPtr<typename Base::SurfaceMesh>   mesh, 
+          const typename Base::SurfaceMesh::Faces& faces, 
+          const Parameters&                        params = parametersShapeGeometry() )
+        {
+          using Face = typename Base::SurfaceMesh::Face;
+
+          bool unit_u = params["unit_u"].as<int>();
+          double radius = params["r-radius"].as<double>();
+          double alpha  = params["alpha"].as<double>();
+          double h      = params["gridstep"].as<double>();
+          if ( alpha != 1.0 ) radius *= pow( h, alpha-1.0 );
+
+          CNCComputer computer(*mesh, unit_u);
+          
+          const auto& mu0  = computer.computeMu0();
+          const auto& muxy = computer.computeMuXY();
+          
+          if (mesh->faceNormals().size() == 0)
+          {
+            // Try to use vertex normals if any
+            if (mesh->vertexNormals().size() == 0)
+              mesh->computeFaceNormalsFromPositions();
+            else 
+              mesh->computeFaceNormalsFromVertexNormals();
+          }
+          
+          const auto& normals = mesh->faceNormals();
+
+          Scalars k1(faces.size()), k2(faces.size());
+          RealVectors d1(faces.size()), d2(faces.size());
+
+          for (size_t i = 0; i < faces.size(); ++i)
+          {
+            const auto center = mesh->faceCentroid(faces[i]);
+            const auto area  = mu0 .measure(center, radius, faces[i]);
+            const auto lmuxy = muxy.measure(center, radius, faces[i]);
+            std::tie(k1[i], k2[i], d1[i], d2[i]) = 
+                CNCComputer::principalCurvatures(area, lmuxy, normals[faces[i]]);
+          }
+
+          return std::make_tuple(k1, k2, d1, d2);
+        }
+
+        /// Given a SurfaceMesh, compute principal curvature at each face using
+        /// CorrectedNormalCurrent method.
+        ///
+        /// This overloads compute curvature for each face of the mesh.
+        ///
+        /// @note If no normals are provided for the faces, the normals will be
+        /// computed (and set) using vertex normals if they exist and positions
+        /// otherwise.
+        ///
+        /// @warning  In this code, only triangles with barycenters strictly
+        /// inside the sphere are considered.
+        ///
+        /// @param[in,out] mesh The surface mesh. The mesh will be modified if
+        /// no face normals are provided.
+        /// @param[in] params
+        ///   - unit_u: Whether the computed normals should be normalized or not
+        ///   - r-radius        [   3.0]: the constant for kernel radius
+        ///   parameter r in r(h)=r h^alpha (VCM,II,Trivial).
+        ///   - alpha           [  0.33]: the parameter alpha in r(h)=r h^alpha
+        ///   (VCM, II, CNC)."
+        ///   - gridstep        [   1.0]: the digitization gridstep (often
+        ///   denoted by h).
+        /// @return The principal curvatures for each face of the mesh, in the
+        /// same order as mesh faces. The result is a 4-element tuples: [first
+        /// curvatures, second curvatures, first directions, second directions].
+        static std::tuple<Scalars, Scalars, RealVectors, RealVectors>
+        getCNCPrincipalCurvaturesAndDirections(
+        CountedPtr<typename Base::SurfaceMesh> mesh,
+        const Parameters & params = parametersShapeGeometry() )
+        {
+          std::vector<typename Base::SurfaceMesh::Face> allFaces(mesh->nbFaces());
+          std::iota(allFaces.begin(), allFaces.end(), 0);
+
+          return getCNCPrincipalCurvaturesAndDirections(mesh, allFaces, params);
+        }
+
+        /// Given a SurfaceMesh, compute principal curvature at each face using
+        /// CorrectedNormalCurrent method.
+        ///
+        /// @warning In this code, only triangles with barycenters strictly
+        /// inside the sphere are considered.
+        ///
+        /// @tparam T Any digital object convertible to surface mesh via
+        /// Shortcuts::makePrimalSurfaceMesh
+        /// @param digitalObject A digital object
+        /// @param params
+        ///   - unit_u: Whether the computed normals should be normalized or not
+        ///   - r-radius        [   3.0]: the constant for kernel radius
+        ///   parameter r in r(h)=r h^alpha (VCM,II,Trivial).
+        ///   - alpha           [  0.33]: the parameter alpha in r(h)=r h^alpha
+        ///   (VCM, II, CNC)."
+        ///   - gridstep        [   1.0]: the digitization gridstep (often
+        ///   denoted by h).
+        /// @return The curvatures for each face of the triangulated surface
+        /// object
+        template <typename T>
+        static std::tuple<Scalars, Scalars, RealVectors, RealVectors>
+        getCNCPrincipalCurvaturesAndDirections(
+        T & digitalObject,
+        const Parameters & params = parametersShapeGeometry() )
+        {
+          CountedPtr<typename Base::SurfaceMesh> mesh = Base::makePrimalSurfaceMesh(digitalObject);
+          return getCNCPrincipalCurvaturesAndDirections(mesh, params);
+        }
+         /// @}
 
       // --------------------------- geometry estimation ------------------------------
       /// @name Geometry estimation services
@@ -639,6 +975,7 @@ namespace DGtal
       ///   - kernel          [ "hat"]: the kernel integration function chi_r, either "hat" or "ball". )
       ///   - alpha           [  0.33]: the parameter alpha in r(h)=r h^alpha (VCM, II)."
       ///   - surfelEmbedding [     0]: the surfel -> point embedding for VCM estimator: 0: Pointels, 1: InnerSpel, 2: OuterSpel.
+      ///   - unit_u [0]: Use unit normals for (CNC) curvature computations.
       static Parameters parametersGeometryEstimation()
       {
         return Parameters
@@ -648,7 +985,8 @@ namespace DGtal
           ( "R-radius",       10.0 )
           ( "r-radius",        3.0 )
           ( "alpha",          0.33 )
-          ( "surfelEmbedding",   0 );
+          ( "surfelEmbedding",   0 )
+          ( "unit_u"         ,   0 );
       }
 
       /// Given a digital space \a K and a vector of \a surfels,
@@ -1286,7 +1624,7 @@ namespace DGtal
       /// @param[in] surfels the sequence of surfels at which we compute the Gaussian curvatures
       /// @param[in] params the parameters:
       ///   - verbose         [     1]: verbose trace mode 0: silent, 1: verbose.
-      ///   - r-radius        [   3.0]: the constant for kernel radius parameter r in r(h)=r h^alpha (VCM,II,Trivial).
+      ///   - r-radius        [   3.0]: the constant for kernel radius parameter r in r(h)=r h^alpha (VCM,II,Trivial).;
       ///   - alpha           [  0.33]: the parameter alpha in r(h)=r h^alpha (VCM, II)."
       ///   - gridstep        [   1.0]: the digitization gridstep (often denoted by h).
       ///